Styling and behavior fixes
Fix many issues with styling related to bemenu and the upwards list
mode. In addition to this, fix/modify some behavior to improve
this(ex: border_size and border_radius are now a double).

+++ client/common/common.h
@@ -21,8 +21,8 @@ struct client {
     uint32_t selected;
     uint32_t monitor;
     uint32_t hmargin_size;
-    uint32_t border_size;
-    uint32_t border_radius;
+    double border_size;
+    double border_radius;
     float width_factor;
     bool bottom;
     bool center;
